feat: add/remove login tab based on credentials existance
